Set up pattern to store hero bands in their own files
