+++ src/framework/types/data/hero.ts
@@ -1,0 +1,9 @@
+import { InstanceCard } from "./card";
+
+export type HeroBand = (id: string) => Partial<InstanceCard>;
+
+export interface HeroBands {
+  topBand: HeroBand;
+  midBand: HeroBand;
+  maxBand: HeroBand;
+}
