+++ src/tests/heroes/captainzane.ts
